rollback file rename
