+++ .eslintrc.js
@@ -54,6 +54,16 @@ module.exports = {
         'website/**/*',
         '**/jest-runtime/**/*',
         '**/src/Console*',
+        'packages/jest-cli/src/lib/Prompt.js',
+        'packages/jest-cli/src/reporters/Status.js',
+        'packages/jest-editor-support/src/Process.js',
+        'packages/jest-editor-support/src/Runner.js',
+        'packages/jest-editor-support/src/Settings.js',
+        'packages/jest-jasmine2/src/jasmine/Env.js',
+        'packages/jest-jasmine2/src/jasmine/Spec.js',
+        'packages/jest-jasmine2/src/jasmine/Suite.js',
+        'packages/jest-jasmine2/src/jasmine/Timer.js',
+        'packages/jest-snapshot/src/State.js',
       ],
       rules: {
         'unicorn/filename-case': 0,
