+++ packages/jest-cli/src/lib/Prompt.js
