+++ packages/jest-cli/src/lib/__tests__/prompt.test.js
@@ -9,7 +9,7 @@
 */
 'use strict';
 
-const Prompt = require('../prompt');
+const Prompt = require('../Prompt');
 let {KEYS} = require('../../constants');
 
 KEYS = Object.assign({}, KEYS, {
