+++ packages/jest-cli/src/lib/pattern_mode_helpers.js
@@ -14,7 +14,7 @@ import chalk from 'chalk';
 import ansiEscapes from 'ansi-escapes';
 import stringLength from 'string-length';
 import stripAnsi from 'strip-ansi';
-import Prompt from './prompt';
+import Prompt from './Prompt';
 
 const pluralize = (count: number, text: string) =>
   count === 1 ? text : text + 's';
