+++ packages/jest-cli/src/reporters/Status.js
