+++ packages/jest-editor-support/src/index.js
@@ -8,11 +8,11 @@
  * @flow
  */
 
-import Process from './process';
+import Process from './Process';
 
 import ProjectWorkspace from './project_workspace';
-import Runner from './runner';
-import Settings from './settings';
+import Runner from './Runner';
+import Settings from './Settings';
 import {Expect, ItBlock, Node} from './parsers/parser_nodes';
 import {parse} from './parsers/babylon_parser';
 import TestReconciler from './test_reconciler';
