+++ packages/jest-jasmine2/src/jasmine/Env.js
