+++ packages/jest-jasmine2/src/jasmine/Suite.js
