+++ packages/jest-jasmine2/src/jasmine/Timer.js
