+++ packages/jest-snapshot/src/State.js
